requirements.txt
streamlit
openai
feedparser
youtube-transcript-api
telethon
python-dotenv
beautifulsoup4
requests
google-generativeai
